+++ Assets/Game/Gameplay/Survival Meсhanics/Scripts/Dehydration/DehydrationConfig.cs
@@ -5,8 +5,8 @@ namespace Game.Gameplay.SurvivalMeсhanics.Dehydration
     [CreateAssetMenu(fileName = "Dehydration Config", menuName = "Configs/Dehydration")]
     public class DehydrationConfig : ScriptableObject
     {
-        [SerializeField, Tooltip("Per Second")] private int _damage;
+        [SerializeField, Tooltip("Per Second")] private float _damage;
         
-        public int Damage => _damage;
+        public float Damage => _damage;
     }
 }
